mon,mgr/pg_autoscaler: tolerate FP rounding in effective_ratio sums
vstart testing surfaced two floating-point issues with the overcommit
checks:

* 0.5 + 0.4 sums to 0.9000000000000001 in binary floating point, so
  creating a third pool pinned at 0.1 was rejected even though the
  operator's pins sum to exactly 1.0 - the primary intended workflow.
  Allow a 1e-6 tolerance in the mon's create/set checks and in the
  mgr's POOL_EFFECTIVE_RATIO_OVERCOMMITTED health check.

* prepare_new_pool() took the ratio as float, so a pin of 0.4 was
  stored as 0.40000000596..., polluting later sums. Carry it as double
  end-to-end like the 'osd pool set' path already did.

+++ src/mon/OSDMonitor.h
@@ -552,7 +552,7 @@ private:
                        uint64_t repl_size,
 		       const uint64_t target_size_bytes,
 		       const float target_size_ratio,
-		       const float effective_ratio,
+		       const double effective_ratio,
 		       const std::string &erasure_code_profile,
                        const unsigned pool_type,
                        const uint64_t expected_num_objects,
